requirements.txt
loguru==0.3.2
Flask==1.1.1
mysql_connector_repackaged==0.3.1
